feat(apps/web): add new logo placement to routes
