+++ apps/web/src/routes/app/index.tsx
@@ -1,6 +1,7 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { PageHeader } from '@/components/app/PageHeader';
 import { SearchInput } from '@/components/search/SearchInput';
+import { PlayrLogo } from '@/components/app/PlayrLogo.tsx';
 
 export const Route = createFileRoute('/app/')({
   component: App,
@@ -9,7 +10,7 @@ export const Route = createFileRoute('/app/')({
 function App() {
   return (
     <div className="flex flex-col gap-4 p-4">
-      <PageHeader title="Home" centerActions>
+      <PageHeader title={<PlayrLogo />} description="Home" centerActions>
         <SearchInput className="max-w-xl" />
       </PageHeader>
     </div>
